Azure: don't use pip to install azure-cli
